requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
decorator==5.2.1
dj-database-url==3.0.1
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
future==1.0.0
geocoder==1.38.1
idna==3.10
psycopg==3.2.9
psycopg-binary==3.2.9
python-dotenv==1.1.1
ratelim==0.1.6
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
wheel==0.45.1
